+++ src/integration/lib/src/test/java/se/tink/backend/aggregation/nxgen/core/account/nxbuilders/modules/loan/LoanModuleTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import com.google.common.collect.Lists;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import org.junit.Test;
 import se.tink.backend.aggregation.nxgen.core.account.loan.LoanDetails.Type;
@@ -20,7 +21,7 @@ public class LoanModuleTest {
     public void nullType() {
         LoanModule.builder()
                 .withType(null)
-                .withBalance(ExactCurrencyAmount.of(2500, "EUR"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(2500), "EUR"))
                 .withInterestRate(0.1)
                 .build();
     }
@@ -34,7 +35,7 @@ public class LoanModuleTest {
     public void negativeInterest() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(2500, "EUR"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(2500), "EUR"))
                 .withInterestRate(-1.24398)
                 .build();
     }
@@ -43,7 +44,7 @@ public class LoanModuleTest {
     public void nullAmortized() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(1_560_000.23, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(1_560_000.23), "SEK"))
                 .withInterestRate(0.01976)
                 .setAmortized(null)
                 .build();
@@ -53,7 +54,7 @@ public class LoanModuleTest {
     public void nullMonthlyAmortization() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(1_560_000.23, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(1_560_000.23), "SEK"))
                 .withInterestRate(0.01976)
                 .setMonthlyAmortization(null)
                 .build();
@@ -63,7 +64,7 @@ public class LoanModuleTest {
     public void nullInitialBalance() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(1_560_000.23, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(1_560_000.23), "SEK"))
                 .withInterestRate(0.01976)
                 .setInitialBalance(null)
                 .build();
@@ -73,7 +74,7 @@ public class LoanModuleTest {
     public void nullInitialDate() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(1_560_000.23, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(1_560_000.23), "SEK"))
                 .withInterestRate(0.01976)
                 .setInitialDate(null)
                 .build();
@@ -83,7 +84,7 @@ public class LoanModuleTest {
     public void nullLoanNumber() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(1_560_000.23, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(1_560_000.23), "SEK"))
                 .withInterestRate(0.01976)
                 .setLoanNumber(null)
                 .build();
@@ -93,7 +94,7 @@ public class LoanModuleTest {
     public void negativeMonthsBound() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(1_560_000.23, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(1_560_000.23), "SEK"))
                 .withInterestRate(0.01976)
                 .setNumMonthsBound(-2)
                 .build();
@@ -103,7 +104,7 @@ public class LoanModuleTest {
     public void nullNextDayOfTermsChange() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(1_560_000.23, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(1_560_000.23), "SEK"))
                 .withInterestRate(0.01976)
                 .setNextDayOfTermsChange(null)
                 .build();
@@ -113,7 +114,7 @@ public class LoanModuleTest {
     public void nullSecurity() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(1_560_000.23, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(1_560_000.23), "SEK"))
                 .withInterestRate(0.01976)
                 .setSecurity(null)
                 .build();
@@ -123,7 +124,7 @@ public class LoanModuleTest {
     public void nullApplicants() {
         LoanModule.builder()
                 .withType(Type.MORTGAGE)
-                .withBalance(ExactCurrencyAmount.of(1_560_000.23, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(1_560_000.23), "SEK"))
                 .withInterestRate(0.01976)
                 .setApplicants(null)
                 .build();
@@ -135,31 +136,41 @@ public class LoanModuleTest {
         LoanModule loanModule =
                 LoanModule.builder()
                         .withType(Type.BLANCO)
-                        .withBalance(ExactCurrencyAmount.of(359_641.23, "SEK"))
+                        .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(359_641.23), "SEK"))
                         .withInterestRate(15.65)
-                        .setAmortized(ExactCurrencyAmount.of(90_358.77, "SEK"))
+                        .setAmortized(ExactCurrencyAmount.of(BigDecimal.valueOf(90_358.77), "SEK"))
                         .setApplicants(Lists.newArrayList("Göran Persson", "Mona Sahlin"))
                         .setCoApplicant(true)
-                        .setInitialBalance(ExactCurrencyAmount.of(450_000, "SEK"))
+                        .setInitialBalance(
+                                ExactCurrencyAmount.of(BigDecimal.valueOf(450_000), "SEK"))
                         .setInitialDate(LocalDate.of(2017, 3, 26))
                         .setLoanNumber("XL47282")
-                        .setMonthlyAmortization(ExactCurrencyAmount.of(5_000.00, "SEK"))
+                        .setMonthlyAmortization(
+                                ExactCurrencyAmount.of(BigDecimal.valueOf(5_000.00), "SEK"))
                         .setNextDayOfTermsChange(LocalDate.of(2022, 5, 1))
                         .setNumMonthsBound(3)
                         .setSecurity("Riksdagshuset")
                         .build();
 
         assertEquals(Type.BLANCO, loanModule.getLoanType());
-        assertEquals(loanModule.getBalance(), ExactCurrencyAmount.of(359_641.23, "SEK"));
+        assertEquals(
+                loanModule.getBalance(),
+                ExactCurrencyAmount.of(BigDecimal.valueOf(359_641.23), "SEK"));
         assertEquals(loanModule.getInterestRate(), 15.65, 0);
-        assertEquals(loanModule.getAmortized(), ExactCurrencyAmount.of(90_358.77, "SEK"));
+        assertEquals(
+                loanModule.getAmortized(),
+                ExactCurrencyAmount.of(BigDecimal.valueOf(90_358.77), "SEK"));
         assertEquals(
                 loanModule.getApplicants(), Lists.newArrayList("Göran Persson", "Mona Sahlin"));
         assertTrue(loanModule.isCoApplicant());
-        assertEquals(loanModule.getInitialBalance(), ExactCurrencyAmount.of(450_000, "SEK"));
+        assertEquals(
+                loanModule.getInitialBalance(),
+                ExactCurrencyAmount.of(BigDecimal.valueOf(450_000), "SEK"));
         assertEquals(loanModule.getInitialDate(), LocalDate.of(2017, 3, 26));
         assertEquals(loanModule.getLoanNumber(), "XL47282");
-        assertEquals(loanModule.getMonthlyAmortization(), ExactCurrencyAmount.of(5_000.00, "SEK"));
+        assertEquals(
+                loanModule.getMonthlyAmortization(),
+                ExactCurrencyAmount.of(BigDecimal.valueOf(5_000.00), "SEK"));
         assertEquals(loanModule.getNextDayOfTermsChange(), LocalDate.of(2022, 5, 1));
         assertEquals(loanModule.getNumMonthsBound(), 3);
         assertEquals(loanModule.getSecurity(), "Riksdagshuset");
